pyproject.toml
[tool.poetry]
name = "n100-tickers"
version = "0.1.0"
description = "Return NASDAQ 100 ticker symbols by date."
authors = ["Jeff McCarrell <jeff@mccarrell.org>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
